classifier implementations rearranged

+++ classification/Decision_Tree_CubeIDE/Inc/tree_inference.h
@@ -1,3 +1,3 @@
 #include "tree_config.h"
 
-int tree_inference(int input[]);
+int tree_inference(float input[]);
